refactor: Updated API and folder structure to match those of @codemod-utils/ast
