--- packages/type-css-modules/src/utils/abstract-syntax-tree.js
@@ -1,1 +1,0 @@
-export { default as ASTCSS } from './abstract-syntax-tree/css.js';
