+++ packages/type-css-modules/src/utils/ast.js
@@ -1,0 +1,1 @@
+export { default as ASTCSS } from './ast/css.js';
