+++ packages/type-css-modules/tests/utils/ast/css/traverse-file-is-empty.test.js
@@ -1,0 +1,11 @@
+import { assert, test } from '@codemod-utils/tests';
+
+import { traverseCSS } from '../../../helpers/transforms/css.js';
+
+test('utils | ast | css > traverse (file is empty)', function () {
+  const oldFile = '';
+
+  const newFile = traverseCSS(oldFile);
+
+  assert.strictEqual(newFile, '');
+});
